Rename data_types modules to snake_case
