+++ data_types/linked_list.py
